# backend/api/swagger.yaml
openapi: 3.0.3
info:
  title: Eiendomsmuligheter Platform API
  description: |-
    Offisiell API-dokumentasjon for Eiendomsmuligheter Platform 2.0
    
    Dette API-et gir tilgang til alle funksjonene i Eiendomsmuligheter Platform, inkludert:
    * Eiendomsanalyse med AlterraML
    * 3D-visualisering med TerravisionJS
    * Kommuneintegrasjon med CommuneConnect
    * Økonomi- og bærekraftsvurderinger
    
    Mer informasjon finnes på [https://eiendomsmuligheter.no](https://eiendomsmuligheter.no)
  contact:
    email: support@eiendomsmuligheter.no
  license:
    name: Proprietær
    url: https://eiendomsmuligheter.no/license
  version: 2.0.0
servers:
  - url: https://api.eiendomsmuligheter.no/api/v1
    description: Produksjonsserver
  - url: https://staging.eiendomsmuligheter.no/api/v1
    description: Testserver
  - url: http://localhost:8000/api
    description: Lokal utviklingsserver
...
paths:
  /property/analyze:
    post:
      tags:
        - property
      summary: Analyser eiendom
      description: Utfører en omfattende analyse av en eiendom basert på adresse og tilleggsdata
      operationId: analyzeProperty
      requestBody:
        description: Eiendomsdata for analyse
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyRequest'
        required: true
      responses:
        '200':
          description: Vellykket analyse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyAnalysisResponse'
        '400':
          description: Ugyldig forespørsel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Serverfeil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
  /municipality/{municipality_id}/regulations:
    get:
      tags:
        - municipality
      summary: Hent alle reguleringsplaner for en kommune
      description: Returnerer en liste over alle tilgjengelige reguleringsplaner for den angitte kommunen
      operationId: getMunicipalityRegulations
      parameters:
        - name: municipality_id
          in: path
          description: Kommunens ID (f.eks. 0301 for Oslo)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vellykket henting av reguleringsplaner
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegulationData'
        '404':
          description: Kommune ikke funnet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Serverfeil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
  /municipality/{municipality_id}/contacts:
    get:
      tags:
        - municipality
      summary: Hent kontaktinformasjon for en kommune
      description: Returnerer kontaktinformasjon for relevante avdelinger i den angitte kommunen
      operationId: getMunicipalityContacts
      parameters:
        - name: municipality_id
          in: path
          description: Kommunens ID (f.eks. 0301 for Oslo)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vellykket henting av kontaktinformasjon
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MunicipalityContact'
        '404':
          description: Kommune ikke funnet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Serverfeil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
...
components:
  schemas:
    PropertyRequest:
      type: object
      required:
        - address
        - lot_size
        - current_utilization
        - building_height
        - floor_area_ratio
      properties:
        property_id:
          type: string
          description: Eiendoms-ID (hvis kjent)
          example: "1234567"
        address:
          type: string
          description: Adressen til eiendommen
          example: "Storgata 1, 0182 Oslo"
        municipality_id:
          type: string
          description: Kommune-ID (hvis kjent)
          example: "0301"
        zoning_category:
          type: string
          description: Reguleringstype
          example: "residential"
        lot_size:
          type: number
          format: float
          description: Tomtestørrelse i kvadratmeter
          example: 500.0
        current_utilization:
          type: number
          format: float
          description: Nåværende utnyttelsesgrad
          example: 0.3
        building_height:
          type: number
          format: float
          description: Bygningshøyde i meter
          example: 12.0
        floor_area_ratio:
          type: number
          format: float
          description: BRA-faktor
          example: 0.7
        images:
          type: array
          items:
            type: string
          description: Liste med bilde-URLs
        additional_data:
          type: object
          additionalProperties: true
          description: Tilleggsdata
    RegulationRule:
      type: object
      required:
        - id
        - rule_type
        - value
        - description
      properties:
        id:
          type: string
          description: Regel-ID
          example: "REG-001"
        rule_type:
          type: string
          description: Regeltype
          example: "floor_area_ratio"
        value:
          type: string
          description: Regelverdi
          example: "0.7"
        description:
          type: string
          description: Beskrivelse
          example: "Maks tillatt BRA faktor"
        unit:
          type: string
          description: Måleenhet
          example: "ratio"
        category:
          type: string
          description: Kategori
          example: "utilization"
    BuildingPotential:
      type: object
      required:
        - max_buildable_area
        - max_height
        - max_units
        - optimal_configuration
      properties:
        max_buildable_area:
          type: number
          format: float
          description: Maksimalt byggbart areal
          example: 350.0
        max_height:
          type: number
          format: float
          description: Maksimal høyde
          example: 12.0
        max_units:
          type: integer
          description: Maksimalt antall enheter
          example: 4
        optimal_configuration:
          type: string
          description: Optimal konfigurasjon
          example: "4 enheter fordelt på 2 etasjer"
        constraints:
          type: array
          items:
            type: string
          description: Begrensninger
          example: ["Avstand til nabo: 4m", "Maks takvinkel: 45°"]
        recommendations:
          type: array
          items:
            type: string
          description: Anbefalinger
          example: ["Plasser bygget mot nord for god solforhold"]
    EnergyProfile:
      type: object
      required:
        - energy_class
        - heating_demand
        - cooling_demand
        - primary_energy_source
      properties:
        energy_class:
          type: string
          description: Energiklasse
          example: "B"
        heating_demand:
          type: number
          format: float
          description: Oppvarmingsbehov
          example: 75.0
        cooling_demand:
          type: number
          format: float
          description: Kjølebehov
          example: 20.0
        primary_energy_source:
          type: string
          description: Primær energikilde
          example: "Electricity"
        recommendations:
          type: array
          items:
            type: string
          description: Anbefalinger
          example: ["Vurder solceller på sørvendt tak"]
    PropertyAnalysisResponse:
      type: object
      required:
        - property_id
        - address
        - regulations
        - building_potential
      properties:
        property_id:
          type: string
          description: Eiendoms-ID
          example: "1234567"
        address:
          type: string
          description: Adresse
          example: "Storgata 1, 0182 Oslo"
        regulations:
          type: array
          items:
            $ref: '#/components/schemas/RegulationRule'
          description: Reguleringsregler
        building_potential:
          $ref: '#/components/schemas/BuildingPotential'
          description: Bygningspotensial
        energy_profile:
          $ref: '#/components/schemas/EnergyProfile'
          description: Energiprofil
        roi_estimate:
          type: number
          format: float
          description: Estimert ROI
          example: 0.15
        risk_assessment:
          type: object
          additionalProperties:
            type: string
          description: Risikovurdering
          example: {"market_risk": "low", "regulatory_risk": "medium"}
        recommendations:
          type: array
          items:
            type: string
          description: Anbefalinger
          example: ["Bygg rekkehus for optimal utnyttelse", "Inkluder grønne tak"]
...
    RegulationData:
      type: object
      required:
        - regulation_id
        - municipality_id
        - title
        - status
        - valid_from
      properties:
        regulation_id:
          type: string
          description: Reguleringsplan-ID
          example: "R-123456"
        municipality_id:
          type: string
          description: Kommune-ID
          example: "0301"
        title:
          type: string
          description: Tittel
          example: "Reguleringsplan for Storgata 1"
        description:
          type: string
          description: Beskrivelse
          example: "Reguleringsplan for boligutvikling i sentrum"
        status:
          type: string
          description: Status
          example: "active"
        valid_from:
          type: string
          format: date
          description: Gyldig fra
          example: "2020-01-01"
        valid_to:
          type: string
          format: date
          description: Gyldig til
          example: "2030-01-01"
        document_url:
          type: string
          description: URL til dokumenter
          example: "https://kommune.no/regulations/R-123456.pdf"
        rules:
          type: array
          items:
            $ref: '#/components/schemas/RegulationRule'
          description: Regler
    MunicipalityContact:
      type: object
      required:
        - municipality_id
        - name
        - department
        - email
        - role
      properties:
        municipality_id:
          type: string
          description: Kommune-ID
          example: "0301"
        name:
          type: string
          description: Navn
          example: "Ola Nordmann"
        department:
          type: string
          description: Avdeling
          example: "Byggesaksavdelingen"
        email:
          type: string
          description: E-post
          example: "ola.nordmann@kommune.no"
        phone:
          type: string
          description: Telefon
          example: "12345678"
        role:
          type: string
          description: Rolle
          example: "Byggesaksbehandler"
...
    ErrorResponse:
      type: object
      required:
        - status_code
        - detail
      properties:
        status_code:
          type: integer
          description: HTTP-statuskode
          example: 400
        detail:
          type: string
          description: Feilbeskrivelse
          example: "Ugyldig forespørsel"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT-autentiseringstoken 